+++ src/lib/components/user-card.svelte
@@ -1,5 +1,6 @@
 <script lang="ts">
 	import type { Prisma, Question } from '@prisma/client';
+	import SvelteMarkdown from 'svelte-markdown';
 
 	export let user: Partial<Prisma.UserGetPayload<{ include: { authUser: true; decision: true } }>>;
 	export let questions: Question[];
@@ -7,13 +8,22 @@
 	$: application = user.application as Record<string, unknown>;
 </script>
 
-<p><b>Role</b> {user.authUser?.role}</p>
+<p><b>Role</b> {user.authUser?.roles.join(', ')}</p>
 <p>
 	<b>Status</b>
 	{user.authUser?.status}
 	{#if user.decision}(Pending {user.decision.status}){/if}
 </p>
-
 {#each questions as question}
-	<p><b>{question.label}</b> {application[question.id]}</p>
+	<SvelteMarkdown source={question.label} />
+	<blockquote>{application[question.id]}</blockquote>
 {/each}
+
+<style>
+	blockquote {
+		background: #eee;
+		padding: 0.5rem;
+		border-left: none;
+		border-left: #aaa 5px solid;
+	}
+</style>