+++ src/lib/trpc/routes/questions.ts
@@ -53,6 +53,76 @@ export const questionsRouter = t.router({
 		}),
 
 	/**
+	 * Moves a question one position down. User must be an admin. Does
+	 * nothing if the question is already at the bottom.
+	 */
+	moveDown: t.procedure
+		.use(authenticate(['ADMIN']))
+		.input(z.string())
+		.mutation(async (req): Promise<void> => {
+			const question = await prisma.question.findUniqueOrThrow({
+				where: { id: req.input },
+			});
+			const nextQuestion = await prisma.question.findFirst({
+				where: { order: { gt: question.order } },
+				orderBy: [{ order: 'asc' }],
+			});
+			if (nextQuestion === null) {
+				return;
+			}
+			// To swap the question orders, we must temporarily set the first
+			// question's order to -1 to avoid a unique constraint violation.
+			const clearX = prisma.question.update({
+				where: { id: question.id },
+				data: { order: -1 },
+			});
+			const setYtoX = prisma.question.update({
+				where: { id: nextQuestion.id },
+				data: { order: question.order },
+			});
+			const setXtoY = prisma.question.update({
+				where: { order: -1 },
+				data: { order: nextQuestion.order },
+			});
+			await prisma.$transaction([clearX, setYtoX, setXtoY]);
+		}),
+
+	/**
+	 * Moves a question one position up. User must be an admin. Does
+	 * nothing if the question is already at the top.
+	 */
+	moveUp: t.procedure
+		.use(authenticate(['ADMIN']))
+		.input(z.string())
+		.mutation(async (req): Promise<void> => {
+			const question = await prisma.question.findUniqueOrThrow({
+				where: { id: req.input },
+			});
+			const prevQuestion = await prisma.question.findFirst({
+				where: { order: { lt: question.order } },
+				orderBy: [{ order: 'desc' }],
+			});
+			if (prevQuestion === null) {
+				return;
+			}
+			// To swap the question orders, we must temporarily set the first
+			// question's order to -1 to avoid a unique constraint violation.
+			const clearX = prisma.question.update({
+				where: { id: question.id },
+				data: { order: -1 },
+			});
+			const setYtoX = prisma.question.update({
+				where: { id: prevQuestion.id },
+				data: { order: question.order },
+			});
+			const setXtoY = prisma.question.update({
+				where: { order: -1 },
+				data: { order: prevQuestion.order },
+			});
+			await prisma.$transaction([clearX, setYtoX, setXtoY]);
+		}),
+
+	/**
 	 * Deletes a question. User must be an admin.
 	 */
 	delete: t.procedure