+++ src/lib/trpc/routes/users.ts
@@ -49,7 +49,7 @@ export const usersRouter = t.router({
 						});
 					}
 				} else {
-					if (user.role !== 'ADMIN') {
+					if (!user.roles.includes('ADMIN')) {
 						throw new Error('Forbidden');
 					}
 					return await prisma.user.findUniqueOrThrow({
@@ -117,7 +117,10 @@ export const usersRouter = t.router({
 			const application = user.application as Record<string, any>;
 			for (const question of questions) {
 				const answer = application[question.id];
-				if (question.required && (answer === undefined || answer === null)) {
+				if (
+					question.required &&
+					(answer === undefined || answer === null || answer === false || answer === '')
+				) {
 					errors[question.label] = 'This field is required.';
 				} else if (
 					(question.type === 'SENTENCE' || question.type === 'PARAGRAPH') &&
@@ -223,16 +226,18 @@ export const usersRouter = t.router({
 		.mutation(async (req): Promise<Session | null> => {
 			try {
 				const user = await auth.createUser({
-					primaryKey: {
-						providerId: 'email',
-						providerUserId: req.input.email,
-						password: req.input.password,
-					},
+					primaryKey: null,
 					attributes: {
 						email: req.input.email,
-						role: 'HACKER',
+						roles: ['HACKER'],
 						status: 'CREATED',
 					},
+				});
+				await auth.createKey(user.id, {
+					type: 'persistent',
+					providerId: 'email',
+					providerUserId: req.input.email,
+					password: req.input.password,
 				});
 				return await auth.createSession(user.id);
 			} catch (e) {
@@ -285,7 +290,7 @@ export const usersRouter = t.router({
 			if (user !== null) {
 				await resetPasswordToken.invalidateAllUserTokens(user.id);
 				const token = await resetPasswordToken.issue(user.id);
-				let link = `${process.env.DOMAIN_NAME}/login/reset-password/${token}`;
+				let link = `${process.env.DOMAIN_NAME}/login/reset-password?token=${token}`;
 				link = `<a href="${link}">${link}</a>`;
 				const body =
 					'Click on the following link to reset your password (valid for 10 minutes):<br><br>' +
@@ -320,7 +325,18 @@ export const usersRouter = t.router({
 			const user = await auth.getUser(token.userId);
 			await auth.invalidateAllUserSessions(user.id);
 			await resetPasswordToken.invalidateAllUserTokens(user.id);
-			await auth.updateKeyPassword('email', user.email, req.input.password);
+			try {
+				await auth.updateKeyPassword('email', user.email, req.input.password);
+			} catch (e) {
+				// If the user doesn't have a password (because they
+				// signed up through a third-party provider), create one
+				await auth.createKey(user.id, {
+					type: 'persistent',
+					providerId: 'email',
+					providerUserId: user.email,
+					password: req.input.password,
+				});
+			}
 			return await auth.createSession(user.id);
 		}),
 
@@ -413,9 +429,9 @@ export const usersRouter = t.router({
 		}),
 
 	/**
-	 * Bulk sets the roles of all the users. User must be an admin.
+	 * Bulk adds a role to all users. User must be an admin.
 	 */
-	setRoles: t.procedure
+	addRole: t.procedure
 		.use(authenticate(['ADMIN']))
 		.input(
 			z.object({
@@ -427,9 +443,50 @@ export const usersRouter = t.router({
 			if (req.input.ids.includes(req.ctx.user.id)) {
 				throw new Error('You cannot change your own role.');
 			}
-			await prisma.authUser.updateMany({
+
+			const users = await prisma.authUser.findMany({
 				where: { id: { in: req.input.ids } },
-				data: { role: req.input.role },
+				select: { id: true, roles: true },
 			});
+
+			for (const user of users) {
+				if (!user.roles.includes(req.input.role)) {
+					const updatedRoles = [...user.roles, req.input.role];
+					await prisma.authUser.update({
+						where: { id: user.id },
+						data: { roles: { set: updatedRoles } },
+					});
+				}
+			}
+		}),
+
+	/**
+	 * Bulk removes a role from all users. The user must be an admin.
+	 */
+	removeRole: t.procedure
+		.use(authenticate(['ADMIN']))
+		.input(
+			z.object({
+				role: z.nativeEnum(Role),
+				ids: z.array(z.string()),
+			})
+		)
+		.mutation(async (req): Promise<void> => {
+			if (req.input.ids.includes(req.ctx.user.id)) {
+				throw new Error('You cannot change your own roles.');
+			}
+
+			const users = await prisma.authUser.findMany({
+				where: { id: { in: req.input.ids } },
+				select: { id: true, roles: true },
+			});
+
+			for (const user of users) {
+				const updatedRoles = user.roles.filter((role) => role !== req.input.role);
+				await prisma.authUser.update({
+					where: { id: user.id },
+					data: { roles: { set: updatedRoles } },
+				});
+			}
 		}),
 });