+++ src/routes/apply/+page.server.ts
@@ -16,20 +16,32 @@ function formToApplication(questions: Question[], formData: FormData) {
 	// eslint-disable-next-line @typescript-eslint/no-explicit-any
 	const application: Record<string, any> = {};
 	for (const question of questions) {
-		if (formData.get(question.id) === '') {
-			application[question.id] = undefined;
-			continue;
-		}
-		if (
-			question.type === 'SENTENCE' ||
-			question.type === 'PARAGRAPH' ||
-			question.type === 'DROPDOWN'
-		) {
+		if (question.type === 'SENTENCE' || question.type === 'PARAGRAPH') {
 			application[question.id] = formData.get(question.id);
 		} else if (question.type === 'NUMBER') {
 			application[question.id] = Number(formData.get(question.id));
 			if (isNaN(application[question.id])) {
 				application[question.id] = undefined;
+			}
+		} else if (question.type === 'CHECKBOX') {
+			application[question.id] = formData.get(question.id) === 'on';
+		} else if (question.type === 'DROPDOWN') {
+			const selected = formData.get(question.id) as string;
+			try {
+				application[question.id] = JSON.parse(selected).value;
+			} catch (e) {
+				// Needed because JSON.parse on an empty string errors
+				console.error(e);
+			}
+		}
+		if (question.type === 'MULTISELECT') {
+			const selected = formData.get(question.id) as string;
+			try {
+				application[question.id] = JSON.parse(selected).map(
+					(item: { value: string }) => item.value
+				);
+			} catch (e) {
+				console.error(e);
 			}
 		}
 	}