+++ src/routes/schedule/+page.svelte
@@ -8,6 +8,7 @@
 	export let form;
 
 	let editedEvent: Event | null = null;
+	// Possibly a HACK: we return an Event if the user has clicked edit and null otherwise
 	$: if (form !== null) {
 		editedEvent = form;
 		if (browser) {
@@ -84,7 +85,7 @@
 			{#if event.start.toDateString() === displayDate.toDateString()}
 				<li class={currentDateTime > event.end ? event.type + ' passed' : event.type}>
 					<!-- Element removal box -->
-					{#if data.user?.role === 'ADMIN'}
+					{#if data.user?.roles.includes('ADMIN')}
 						<div class="modification-buttons">
 							<form method="POST" use:enhance>
 								<input type="hidden" name="id" value={event.id} />
@@ -115,7 +116,7 @@
 		{/each}
 	</ul>
 </div>
-{#if data.user?.role === 'ADMIN'}
+{#if data.user?.roles.includes('ADMIN')}
 	<hr />
 	<h2>{editedEvent == null ? 'Create New Event' : 'Edit Event'}</h2>
 	<form method="POST" action={editedEvent == null ? '?/create' : '?/saveEdit'} use:enhance>