+++ src/routes/unverified/+page.server.ts
@@ -1,4 +1,12 @@
 import { trpc } from '$lib/trpc/router';
+import { redirect } from '@sveltejs/kit';
+
+export const load = async ({ locals }) => {
+	const { user } = await locals.auth.validateUser();
+	if (user == null || user.status !== 'CREATED') {
+		throw redirect(303, '/');
+	}
+};
 
 export const actions = {
 	default: async ({ locals }) => {